+++ test-framework/core/src/main/java/org/keycloak/testframework/realm/ClientConfigBuilder.java
@@ -34,6 +34,11 @@ public class ClientConfigBuilder {
         return this;
     }
 
+    public ClientConfigBuilder id(String id) {
+        rep.setId(id);
+        return this;
+    }
+
     public ClientConfigBuilder secret(String secret) {
         rep.setSecret(secret);
         return this;
@@ -41,6 +46,11 @@ public class ClientConfigBuilder {
 
     public ClientConfigBuilder name(String name) {
         rep.setName(name);
+        return this;
+    }
+
+    public ClientConfigBuilder description(String description) {
+        rep.setDescription(description);
         return this;
     }
 
@@ -56,6 +66,11 @@ public class ClientConfigBuilder {
 
     public ClientConfigBuilder adminUrl(String adminUrl) {
         rep.setAdminUrl(adminUrl);
+        return this;
+    }
+
+    public ClientConfigBuilder baseUrl(String baseUrl) {
+        rep.setBaseUrl(baseUrl);
         return this;
     }
 
@@ -82,6 +97,11 @@ public class ClientConfigBuilder {
 
     public ClientConfigBuilder authorizationServices() {
         rep.setAuthorizationServicesEnabled(true);
+        return this;
+    }
+
+    public ClientConfigBuilder fullScopeEnabled(boolean enabled) {
+        rep.setFullScopeAllowed(enabled);
         return this;
     }
